+++ src/main/java/racingcar/dao/PlayerResultDao.java
@@ -7,7 +7,6 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
-import racingcar.service.PlayerResult;
 
 @Repository
 public class PlayerResultDao {
